+++ stelvio/aws/acm.py
@@ -7,11 +7,7 @@ import pulumi_aws
 
 from stelvio import context
 from stelvio.component import Component
-from stelvio.dns import (
-    DnsProviderNotConfiguredError,
-    Record,
-    _call_with_optional_resource_options,
-)
+from stelvio.dns import DnsProviderNotConfiguredError, Record
 from stelvio.provider import ProviderStore
 
 if TYPE_CHECKING:
@@ -86,8 +82,7 @@ class AcmValidatedDomain(
 
         # 2 - Validate Certificate with DNS PROVIDER
         first_option = certificate.domain_validation_options.apply(lambda options: options[0])
-        validation_record = _call_with_optional_resource_options(
-            dns.create_caa_record,
+        validation_record = dns.create_caa_record(
             resource_name=context().prefix(f"{self.name}-certificate-validation-record"),
             name=first_option.apply(lambda opt: opt["resource_record_name"]),
             **self._customizer(